Refactor folder,model and interfaces names
