+++ src/types/requestWIthUser.ts
@@ -1,4 +1,3 @@
-import { UserPayloadDecodedFromToken } from "./UserPayloadDecodedFromToken";
 import { Request } from "express";
 
 export interface RequestWithUser extends Request {
